Move `dask-core` next to `dask`
Adding `dask-core` is superfluous as we already require `dask`. However
the bot will keep adding it on version updates. So we might as keep it
in a way we are comfortable with. When it is next to `dask`, we are more
likely to make sure the 2 have versions that are in sync when the bot
changes one.